==> DNNREC_ML1M/2. Leverage Side Information.ipynb <==
import pickle

PATH='/home/kirana/Documents/phd/experiment1_new/DNNREC_ML1M'
DATAPATH='/home/kirana/Documents/final_dissertation_final/experiments/datasets/ml-1m'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy as sp
import sklearn as sklearn
import datetime as dt 
import seaborn as sns
import torch as torch
from sklearn import metrics
import pickle

[dftrain,dfvalid]=pickle.load(open(f'{DATAPATH}/df.pkl','rb'))

dftrain.shape, dfvalid.shape

dftrain.columns

item_features=['genre','title']
user_features=['gender','occupation','zip']

dftrain[item_features].head(),dftrain[user_features].head()

dftrain['genres']=[item.replace("|", " ") for item in dftrain['genre']]
dfvalid['genres']=[item.replace("|", " ") for item in dfvalid['genre']]

from sklearn import feature_extraction, preprocessing
myvec_genre=sklearn.feature_extraction.text.CountVectorizer (ngram_range=(1,1),lowercase=True)
one_hot_genre_train=myvec_genre.fit_transform(dftrain['genres'])
one_hot_genre_valid=myvec_genre.transform(dfvalid['genres'])

from sklearn import feature_extraction, preprocessing
myvec_title=sklearn.feature_extraction.text.CountVectorizer (ngram_range=(1,1),lowercase=True,max_features=250)
one_hot_title_train=myvec_title.fit_transform(dftrain['title'])
one_hot_title_valid=myvec_title.transform(dfvalid['title'])

from sklearn import feature_extraction, preprocessing
myvec_tag=sklearn.feature_extraction.text.CountVectorizer (ngram_range=(1,1),lowercase=True,max_features=100)
one_hot_tag_train=myvec_tag.fit_transform(dftrain['title'])
one_hot_tag_valid=myvec_tag.transform(dfvalid['title'])

one_hot_title_train.shape, one_hot_genre_train.shape,one_hot_tag_train.shape

one_hot_title_train

df_sideinfo_title_train=pd.DataFrame(one_hot_title_train.todense(),columns=['f_title_'+x for x in myvec_title.get_feature_names()])
df_sideinfo_genre_train=pd.DataFrame(one_hot_genre_train.todense(),columns=['f_genre_'+x for x in myvec_genre.get_feature_names()])
df_sideinfo_tag_train=pd.DataFrame(one_hot_tag_train.todense(),columns=['f_tag_'+x for x in myvec_tag.get_feature_names()])

df_sideinfo_title_valid=pd.DataFrame(one_hot_title_valid.todense(),columns=['f_title_'+x for x in myvec_title.get_feature_names()])
df_sideinfo_genre_valid=pd.DataFrame(one_hot_genre_valid.todense(),columns=['f_genre_'+x for x in myvec_genre.get_feature_names()])
df_sideinfo_tag_valid=pd.DataFrame(one_hot_tag_valid.todense(),columns=['f_tag_'+x for x in myvec_tag.get_feature_names()])

df_sideinfo_title_train.shape,df_sideinfo_genre_train.shape,df_sideinfo_tag_train.shape

df_sideinfo_title_valid.shape,df_sideinfo_genre_valid.shape,df_sideinfo_tag_valid.shape

dftrain.shape

dftrain.dtypes

dftrain.shape,df_sideinfo_title_train.shape,df_sideinfo_genre_train.shape,df_sideinfo_tag_train.shape

dftrain.reset_index(inplace=True)
dfvalid.reset_index(inplace=True)

dftrain=pd.concat([dftrain,df_sideinfo_title_train,df_sideinfo_genre_train,df_sideinfo_tag_train],axis=1)

dfvalid=pd.concat([dfvalid,df_sideinfo_title_valid,df_sideinfo_genre_valid,df_sideinfo_tag_valid],axis=1)

dftrain.shape,dfvalid.shape

dftrain.dtypes

DATAPATH

pickle.dump([dftrain,dfvalid],open(f'{DATAPATH}/df_side.pkl','wb'))
