==> DNNREC_ML1M/3. Assess Cold Start.ipynb <==
import pickle

PATH='/home/kirana/Documents/phd/experiment1_new/DNNREC_ML1M'
DATAPATH='/home/kirana/Documents/final_dissertation_final/experiments/datasets/ml-1m'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy as sp
import sklearn as sklearn
import datetime as dt 
import seaborn as sns
import torch as torch
from sklearn import metrics
import pickle

[dftrain,dfvalid]=pickle.load(open(f'{DATAPATH}/df_side.pkl','rb'))

trainusers=set(dftrain['userId'].unique())
validusers=set(dfvalid['userId'].unique())
trainitems=set(dftrain['itemId'].unique())
validitems=set(dfvalid['itemId'].unique())

commonusers=trainusers.intersection(validusers)
commonitems=trainitems.intersection(validitems)

dfvalid['cold_start_user']=~dfvalid['userId'].isin(commonusers)
dfvalid['cold_start_item']=~dfvalid['itemId'].isin(commonitems)

print (pd.crosstab(dfvalid['cold_start_user'],dfvalid['cold_start_item']))
print (pd.crosstab(dfvalid['cold_start_user'],dfvalid['cold_start_item'])/dfvalid.shape[0])   

dfvalid['cold_start_group']=np.where(dfvalid['cold_start_user'] & dfvalid['cold_start_item'],'User-Item-cold-start',
                                    np.where((~dfvalid['cold_start_user']) & (~dfvalid['cold_start_item']), 'No-cold-start',\
                                            np.where(dfvalid['cold_start_user'], 'User-Cold-Start', 'Item-Cold_start')))    


dfvalid['cold_start_bucket']=np.where(dfvalid['cold_start_group']=='No-cold-start',0,1)

pickle.dump([dftrain,dfvalid],open(f'{DATAPATH}/df_side_cold.pkl','wb'))
