==> DNNREC_ML1M/4. Try different methods.ipynb <==
import pickle

PATH='/home/kirana/Documents/phd/experiment1_new/DNNREC_ML1M'
DATAPATH='/home/kirana/Documents/final_dissertation_final/experiments/datasets/ml-1m'

MYMEDIALITEPATH='/home/kirana/MyMediaLite/bin'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy as sp
import sklearn as sklearn
import datetime as dt 
import seaborn as sns
import torch as torch

from sklearn import metrics

[dftrain,dfvalid]=pickle.load(open(f'{DATAPATH}/df_side_cold.pkl','rb'))

dftrain['rating'].describe()

dfinp_train=dftrain[['userId','itemId','rating']]
dfinp_test=dfvalid[['userId','itemId','rating']]
dfinp_train.to_csv(f'{PATH}/inp_train_ml1m.csv',index=False,header=False,sep=",")
dfinp_test.to_csv(f'{PATH}/inp_test_ml1m.csv',index=False,header=False,sep=",")

TRAINFILEPATH=f'{PATH}/inp_train_ml1m.csv'
TESTFILEPATH=f'{PATH}/inp_test_ml1m.csv'

dfinp_train.head()

### Global Average

cmd=f'{MYMEDIALITEPATH}/rating_prediction --training-file={TRAINFILEPATH} --recommender=GlobalAverage \
--test-file={TESTFILEPATH} --prediction-file {DATAPATH}/prediction_globalavg.txt'

!command {cmd}

predictions_globalavg=pd.read_csv (f'{DATAPATH}/prediction_globalavg.txt',sep='\t',header=None).iloc[:,2]

dfvalid['pred_globalavg']=predictions_globalavg

def score_by_group (dfvalid,predVar,groupVar='cold_start_group',method='global_avg'):
    scores=[]
    mae_scores=[]
    r2_scores=[]
    groups=dfvalid[groupVar].value_counts().index.tolist()
    for group in groups:
        df=dfvalid.loc[dfvalid[groupVar]==group]
        actual=df['rating']
        score=metrics.mean_squared_error(df['rating'],df[predVar])
        mae_score=metrics.mean_absolute_error(df['rating'],df[predVar])
        r2_score=metrics.r2_score(df['rating'],df[predVar])
        scores.append(score)
        mae_scores.append(mae_score)
        r2_scores.append(r2_score)
    return pd.DataFrame({'method':method, 'mse':scores, 'rmse':np.sqrt(scores), 'mae':mae_scores, 'r2_score':r2_scores, 'Cold Start Group':groups})

result_globalavg=pd.concat([score_by_group(dfvalid,'pred_globalavg'),score_by_group(dfvalid,'pred_globalavg','cold_start_bucket'),\
score_by_group(dfvalid,'pred_globalavg','flag_train')],axis=0)

result_globalavg

dfvalid['flag_train'].value_counts()

### User Average

cmd=f'{MYMEDIALITEPATH}/rating_prediction --training-file={TRAINFILEPATH} --recommender=UserAverage \
--test-file={TESTFILEPATH} --prediction-file {DATAPATH}/prediction_useravg.txt'

!command {cmd}

predictions_useravg=pd.read_csv (f'{DATAPATH}/prediction_useravg.txt',sep='\t',header=None).iloc[:,2]

dfvalid['pred_useravg']=predictions_useravg

result_useravg=pd.concat([score_by_group(dfvalid,'pred_useravg','cold_start_group','useravg'),score_by_group(dfvalid,'pred_useravg','cold_start_bucket','useravg'),\
score_by_group(dfvalid,'pred_useravg','flag_train','useravg')],axis=0)
result_useravg

### Item Average

cmd=f'{MYMEDIALITEPATH}/rating_prediction --training-file={TRAINFILEPATH} --recommender=ItemAverage \
--test-file={TESTFILEPATH} --prediction-file {DATAPATH}/prediction_itemavg.txt'

!command {cmd}

predictions_itemavg=pd.read_csv (f'{DATAPATH}/prediction_itemavg.txt',sep='\t',header=None).iloc[:,2]

dfvalid['pred_itemavg']=predictions_itemavg

result_itemavg=pd.concat([score_by_group(dfvalid,'pred_itemavg','cold_start_group','itemavg'),score_by_group(dfvalid,'pred_itemavg','cold_start_bucket','itemavg'),\
score_by_group(dfvalid,'pred_itemavg','flag_train','itemavg')],axis=0)
result_itemavg

## Neighborhood Methods

### Item KNN Cosine

cmd=f'{MYMEDIALITEPATH}/rating_prediction --training-file={TRAINFILEPATH} --recommender=ItemKNN \
--test-file={TESTFILEPATH} --prediction-file {DATAPATH}/prediction_itemknn_cosine.txt \
--recommender-options k=80  --recommender-options reg_u=25 \
--recommender-options reg_i=10'

!command {cmd}

predictions_itemknn_cosine=pd.read_csv (f'{DATAPATH}/prediction_itemknn_cosine.txt',sep='\t',header=None).iloc[:,2]
dfvalid['pred_itemknn_cosine']=predictions_itemknn_cosine

result_itemknncosine=pd.concat([score_by_group(dfvalid,'pred_itemknn_cosine','cold_start_group','itemknncosine'),score_by_group(dfvalid,'pred_itemknn_cosine','cold_start_bucket','itemknncosine'),\
score_by_group(dfvalid,'pred_itemknn_cosine','flag_train','itemknncosine')],axis=0)
result_itemknncosine

### Item KNN Pearson

cmd=f'{MYMEDIALITEPATH}/rating_prediction --training-file={TRAINFILEPATH} --recommender=ItemKNN \
--test-file={TESTFILEPATH} --prediction-file {DATAPATH}/prediction_itemknn_pearson.txt \
--recommender-options k=80  --recommender-options reg_u=25 \
--recommender-options reg_i=10 --recommender-options correlation=Pearson'

!command {cmd}

predictions_itemknn_pearson=pd.read_csv (f'{DATAPATH}/prediction_itemknn_pearson.txt',sep='\t',header=None).iloc[:,2]
dfvalid['pred_itemknn_pearson']=predictions_itemknn_pearson

result_itemknnpearson=pd.concat([score_by_group(dfvalid,'pred_itemknn_pearson','cold_start_group','itemknnpearson'),score_by_group(dfvalid,'pred_itemknn_pearson','cold_start_bucket','itemknnpearson'),\
score_by_group(dfvalid,'pred_itemknn_pearson','flag_train','itemknnpearson')],axis=0)
result_itemknnpearson

### User KNN Pearson

cmd=f'{MYMEDIALITEPATH}/rating_prediction --training-file={TRAINFILEPATH} --recommender=UserKNN \
--test-file={TESTFILEPATH} --prediction-file {DATAPATH}/prediction_userknn_pearson.txt \
--recommender-options k=80  --recommender-options reg_u=25 \
--recommender-options reg_i=10 --recommender-options correlation=Pearson'

!command {cmd}

predictions_userknn_pearson=pd.read_csv (f'{DATAPATH}/prediction_userknn_pearson.txt',sep='\t',header=None).iloc[:,2]
dfvalid['pred_userknn_pearson']=predictions_userknn_pearson

result_userknnpearson=pd.concat([score_by_group(dfvalid,'pred_userknn_pearson','cold_start_group','userknnpearson'),score_by_group(dfvalid,'pred_userknn_pearson','cold_start_bucket','userknnpearson'),\
score_by_group(dfvalid,'pred_userknn_pearson','flag_train','userknnpearson')],axis=0)
result_userknnpearson

### User KNN Cosine

cmd=f'{MYMEDIALITEPATH}/rating_prediction --training-file={TRAINFILEPATH} --recommender=UserKNN \
--test-file={TESTFILEPATH} --prediction-file {DATAPATH}/prediction_userknn_cosine.txt \
--recommender-options k=80  --recommender-options reg_u=25 \
--recommender-options reg_i=10'

!command {cmd}

predictions_userknn_cosine=pd.read_csv (f'{DATAPATH}/prediction_userknn_cosine.txt',sep='\t',header=None).iloc[:,2]
dfvalid['pred_userknn_cosine']=predictions_userknn_cosine

result_userknncosine=pd.concat([score_by_group(dfvalid,'pred_userknn_cosine','cold_start_group','userknncosine'),score_by_group(dfvalid,'pred_userknn_cosine','cold_start_bucket','userknncosine'),\
score_by_group(dfvalid,'pred_userknn_cosine','flag_train','userknncosine')],axis=0)
result_userknncosine

## Matrix Factorization

### SVD

cmd=f'{MYMEDIALITEPATH}/rating_prediction --training-file={TRAINFILEPATH} --recommender=MatrixFactorization \
--test-file={TESTFILEPATH} --prediction-file {DATAPATH}/prediction_svd.txt \
--recommender-options num_factors=10  --recommender-options regularization=0.05 \
--recommender-options learn_rate=0.005 --recommender-options num_iter=75'

!command {cmd}

predictions_svd=pd.read_csv (f'{DATAPATH}/prediction_svd.txt',sep='\t',header=None).iloc[:,2]
dfvalid['pred_svd']=predictions_svd

result_svd=pd.concat([score_by_group(dfvalid,'pred_svd','cold_start_group','svd'),score_by_group(dfvalid,'pred_svd','cold_start_bucket','svd'),\
score_by_group(dfvalid,'pred_svd','flag_train','svd')],axis=0)
result_svd

### SVD Plus Plus

cmd=f'{MYMEDIALITEPATH}/rating_prediction --training-file={TRAINFILEPATH} --recommender=SVDPlusPlus \
--test-file={TESTFILEPATH} --prediction-file {DATAPATH}/prediction_svdplusplus.txt \
--recommender-options num_factors=20  --recommender-options regularization=0.05 \
--recommender-options learn_rate=0.005 --recommender-options num_iter=80'



cmd

!command {cmd}

predictions_svdplusplus=pd.read_csv (f'{DATAPATH}/prediction_svdplusplus.txt',sep='\t',header=None).iloc[:,2]
dfvalid['pred_svdplusplus']=predictions_svdplusplus

result_svdplusplus=pd.concat([score_by_group(dfvalid,'pred_svdplusplus','cold_start_group','svdplusplus'),score_by_group(dfvalid,'pred_svdplusplus','cold_start_bucket','svdplusplus'),\
score_by_group(dfvalid,'pred_svdplusplus','flag_train','svdplusplus')],axis=0)
result_svdplusplus

### BiasedMatrixFactorization

cmd=f'{MYMEDIALITEPATH}/rating_prediction --training-file={TRAINFILEPATH} --recommender=BiasedMatrixFactorization \
--test-file={TESTFILEPATH} --prediction-file {DATAPATH}/prediction_biasedmatrixfactorization.txt \
--recommender-options num_factors=40   --recommender-options bias_reg=0.001 \
--recommender-options learn_rate=0.07 --recommender-options num_iter=110 \
--recommender-options bold_driver=True \
--recommender-options regularization=0.06'

!command {cmd}

predictions_biasedmatrixfactorization=pd.read_csv (f'{DATAPATH}/prediction_biasedmatrixfactorization.txt',sep='\t',header=None).iloc[:,2]
dfvalid['pred_biasedmatrixfactorization']=predictions_biasedmatrixfactorization

result_biasedmatrixfactorization=pd.concat([score_by_group(dfvalid,'pred_biasedmatrixfactorization','cold_start_group','biasedmatrixfactorization'),score_by_group(dfvalid,'pred_biasedmatrixfactorization','cold_start_bucket','biasedmatrixfactorization'),\
score_by_group(dfvalid,'pred_biasedmatrixfactorization','flag_train','biasedmatrixfactorization')],axis=0)
result_biasedmatrixfactorization

results_concat=pd.concat ([result_biasedmatrixfactorization,
result_svdplusplus,
result_svd,
result_userknncosine,
result_userknnpearson,
result_itemknnpearson,
result_itemknncosine,
result_itemavg,
result_useravg,
result_globalavg],axis=0)

pickle.dump(results_concat,open(f'{DATAPATH}/diff_results.pkl','wb'))

results_concat[:25]

pickle.dump([dftrain,dfvalid,results_concat],open(f'{DATAPATH}/df_side_cold_runother.pkl','wb'))

results_concat.sort_index(by='rmse')
